A somewhat working entity system, with various previously working systems now broken.

+++ diamond_rush/src/wall.cpp
@@ -1,8 +1,14 @@
 #include "wall.h"
 
 
-Wall::Wall(int x,int y,SDL_Renderer* renderer) {
-	rect = {x,y,SPRITE_SIZE,SPRITE_SIZE};
+Wall::Wall(float x, float y, SDL_Renderer* renderer) {
+	
+	rect = { 
+		x, y,
+		SPRITE_SIZE, SPRITE_SIZE
+	};
+
 	texture = IMG_LoadTexture(renderer, "data/sprite_wall.png");
+	
 	type = "wall";
 }
